refactor: rename tenderdashInitTask
